refactor delete, restore and forceDelete to reduce duplicated code
